Cristian/sink tweets to database (#27)
* Encrypted DB username using travis and added env variable

* Encrypted database password using travis

* Added rough feature to sink tweets in remote database

* Added a different mapping to make API call and sink to db

* Implemented services to sink tweets in RDS database

* Completed final touches for sink service and adjusted location of some classes / packages

* Adjusted code to ensure all tests pass

* More tests changes and added new sonar exclusions for methods calling the JPA interface. No need to test JAP :)

* Added mockito, Autowired constructors instead of fields and added new return values

+++ src/main/java/com/codingnomads/betty/logic/services/ProbabilityToWinService.java
@@ -34,7 +34,7 @@ public class ProbabilityToWinService {
             throw new InvalidScoreException("Team Score cannot be negative");
         }
 
-        return (double) (teamScore/100);
+        return teamScore/100.0;
     }
